refactor: added useeffect with local storage
