+++ src/utils/constant.js
@@ -3,5 +3,4 @@ export const QUIZ_ACTIONS = {
   SET_USER_SELECTED_OPTION: "SET_USER_SELECTED_OPTION",
   INCREASE_SCORE: "INCREASE_SCORE",
   RESET_QUIZ_STATE: "RESET_QUIZ_STATE",
-  ADD_SCORE_TO_QUIZBOARD: "ADD_SCORE_TO_QUIZBOARD",
 };
